+++ src/module/data/item/ability.mjs
@@ -137,6 +137,11 @@ export default class AbilityModel extends BaseItemModel {
         }
       }
     }
+
+    // Prepare PRE data that relies on ability data prep being complete (e.g. treasure damage bonuses).
+    for (const effect of this.power.effects) {
+      effect.preparePostAbilityPrepData();
+    }
   }
 
   /* -------------------------------------------------- */
@@ -194,7 +199,8 @@ export default class AbilityModel extends BaseItemModel {
         }
 
         if (applyBonus) {
-          const field = DamagePowerRollEffect.schema.getField(bonus.key);
+          // TODO: Remove in v14 with non-persisted schema fields.
+          const field = (bonus.key.startsWith("damage.bonuses")) ? new fields.NumberField({ integer: true }) : DamagePowerRollEffect.schema.getField(bonus.key);
           const firstDamageEffect = this.power.effects.find(effect => effect.type === "damage");
           if (!firstDamageEffect) return;
           const currentValue = foundry.utils.getProperty(firstDamageEffect, bonus.key);